test email changes (#4)
The below message is copied from chapel-lang.

This is to test the log formatting.

, it seems that when we initialize a variable whose
type is `: domain` (currently, the completely generic domain type
specifier), we don't do anything to verify that the initializing
expression actually is a domain, and the variable ends up simply taking
on its type. For example, `var d: domain = 3.14;` would compile and `d`
would have type `real`.

From the little bit of testing I've done, this seems like a gap that's
specific to `domain`, which is to say, I'm not seeing similar problems
for generic classes or records. This is likely due to its special
handling in the compiler. It's also somewhat of an effect of our not
(yet, completely) using copy initializers for domains.

Here, I've addressed this by adding a check to a part of the resolution
pass that's reasoning about source/destination types in initializers—and
domain destination types in particular—to check and make sure the RHS
expression is a domain as well, and to generate an error if not. Happily
this didn't trip over any code in our test system that was relying on
the loophole that's existed.

Resolves #21537.