Final testing.. test formatting and diff (#17)
copied message from chapel-lang to test formatting.

This PR fixes a bug in gen_release that would prevent building
from a non-local repo. There are some other fixes for carrying the
git SHA forward from the source when using gen_release to build
from a remote repo.

The major change is and update to gen_release to explicitly copy
the generated files that contain the git SHA, git-version.cpp and
BUILD_VERSION. This is to fix a bug that always forced the dummy
SHA if not using the local git repo with CHPL_GEN_RELEASE_NO_CLONE
This bug was introduced in chapel-lang/chapel#20920.

The first deviation from previous behavior in gen_release is that we will include the
git SHA in the tarball regardless of how gen_release was called. This will place
all of the decision of whether to print a git SHA from --version or not with
the status of the official_release flag.

The second deviation from previous behavior is that the output of --version
will no longer depend on the developer flag with official releases. The behavior
will be to always print the git SHA if we are not in an official release,
as indicated by the official_release flag, and never print the SHA if the flag is set.

Additional discussion on Cray/chapel-private#4285 also led
to a decision to stop printing the git SHA when the official_release flag is set
and those changes are included here as well.

TESTING:

paratest [Summary: #Successes = 14452 | #Failures = 0 | #Futures = 898]
gen_release 1.30 (SHA exists, but only printed if offical_release is false when building from the tarball)
gen_release (SHA exists and is only printed if official_release if false when building from the tarball)
reviewed by @mppf - thank you!

Signed-off-by: bhavanijayakumaran 82669529+bhavanijayakumaran@users.noreply.github.com